[RFC] Update cached pwm values on pwm store
Userspace application fan2go expects written pwm value to be returned on
read after 5 ms. If the value read back then fan2go assumes that the pwm
is resolution limited and was clamped by the driver.

The nct6687d driver returns cached pwm values on reads. The cache is
updated at most once per second (unless resumed from suspend) and
therefore it leads for fan2go to falsely analyzing the pwm output as
resolution limited.

Avoid the issue by waiting for pwm value to be reported back on each
write and updating the cache immediately. On MSI PRO Z790-P WIFI
(MS-7E06) it takes around 500 ms for the pwm value (with DC mode fan
connected, so it is not a real PWM but a voltage) to be read back after
write.

Signed-off-by: Tomasz Moń <desowin@gmail.com>