Makefile: don't install the file as is for DKMS
Using the whole Makefile results in DKMS running commands unrelated to
the build process (like git commands) which only produce confusing error
output.

Just write `obj-m += nct6687.o` to the Makefile for DKMS as that's the
only thing necessary for it to work.

This also gets rid of `MAKE[0]` in `dkms.conf`.